第二組_嘉義火雞肉飯Dotch_QA/requirements.txt
ollama
gradio
numpy
PyPDF
python-dotenv
google-generativeai
langchain-community
langchain
langchain-ollama
langchain_chroma
langchain-google-genai
bs4
